+++ tests/formatter/test_tox_section.py
@@ -6,7 +6,7 @@ from textwrap import dedent
 import pytest
 
 from tox_ini_fmt.formatter import format_tox_ini
-from tox_ini_fmt.formatter.tox_section import order_env_list
+from tox_ini_fmt.formatter.util import order_env_list
 
 
 def test_no_tox_section(tox_ini: Path) -> None:
