+++ tox.ini
@@ -56,6 +56,7 @@ commands =
 description = check that the long description is valid
 package = skip
 deps =
+
     build[virtualenv]>=0.10
     twine>=4.0.2
 commands =
